No longer delete very recent staff created GEIQ/EA

+++ itou/siaes/management/commands/_import_siae/utils.py
@@ -7,6 +7,8 @@ import os
 from functools import wraps
 from time import time
 
+from django.utils import timezone
+
 from itou.siaes.models import Siae
 from itou.utils.address.models import AddressMixin
 from itou.utils.apis.geocoding import get_geocoding_data
@@ -151,11 +153,18 @@ def sync_structures(df, name, kinds, build_structure, dry_run):
     print(f"{len(deletable_sirets)} {name} will be deleted.")
     for siret in deletable_sirets:
         siae = Siae.objects.get(siret=siret, kind__in=kinds)
+
+        one_week_ago = timezone.now() - timezone.timedelta(days=7)
+        if siae.source == Siae.SOURCE_STAFF_CREATED and siae.created_at >= one_week_ago:
+            print(f"siae.id={siae.id} is staff created and too recent to be deleted.")
+            continue
+
         if could_siae_be_deleted(siae):
             print(f"siae.id={siae.id} will be deleted.")
             if not dry_run:
                 siae.delete()
-        else:
-            # As of 2020/10/16, 5 GEIQ are undeletable.
-            # As of 2020/12/11, 5 EA and EATT are undeletable.
-            print(f"siae.id={siae.id} cannot be deleted as it has data.")
+            continue
+
+        # As of 2020/10/16, 5 GEIQ are undeletable.
+        # As of 2020/12/11, 5 EA and EATT are undeletable.
+        print(f"siae.id={siae.id} cannot be deleted as it has data.")
